fix: TRLST_255 traceback when org is null.
Recent refactoring of trade_remedies_api/core/services/base.py base view introduced a small issue where, when the view is invoked, in certain circumstances the organisation is null and referencing its id causes a traceback. This was spotted in the logs but not visible to the user on the UI (the error seems to be dutifully ignored and the API responds to the client normally, however this could cause issues when called from other pages in the Public and Caseworker portal).