fix: trlst 259 issue post tasks
Investigating an issue under TRLST 259 found some non critical problems in TR API. This PR is mainly a tidy up of the base settings file and the local.env.example file. also bumped the Postgres version used for local development:
- Removed environ.Env.read_env(). We use docker-compose for local development
and feed the container settings from a local.env file. In PaaS we pick up the
app's environment and do not deploy a '.env' file.
- Simplified REDIS URL definition and removed hard coding of database number.
- Simplified CELERY_BROKER_URL definition and removed database number hardcoding.
- Removed 'AWS_DEFAULT_REGION' this is an AWS API client setting and not used
in the codebase.
- Corrected the typo in S3_DOWNLOAD_LINK_EXPIRY_SECONDS var name.
- Removed 'TRUSTED_USER_TOKEN' setting, this is not used in the code base.
- Removed unused imports.
- Removed cruft comments.
- Pepified.