feat: trlst 219 extend 2fa period
The requirement for this change is to make it possible to extend the Two-Factor Authentication code validity period, but just for the email delivery type. Users in certain regions were experiencing severe lag in 2FA code deliver (via SMS and Email). In lieu of adding an app based authenticator solution (lots of work) special dispensation has been given by cyber to extend the 2FA Code validity period but just for email delivery as sms delivery is deemed less secure and more prone to interception. Summary of changes:
-  Added delivery_type to the TwoFactorAuth model, limited to DELIVERY_TYPE_CHOICES. This enables validation of the code's expiry based on delivery type (currently email or sms).
- Added new setting for 2FA code validity for emails.
- General tidy up, improvement of setting and var names, doc strings.