fix: more specific catch DoesNotExist.
Note the entire post method is crazy and should be refactored, unfortunately it's pivotal in all document operations. This is a minimal fix pending deeper analysis of code.

Fixes an issue in TR API where, when a child document does not exist, or does not exist for a submission `DoesNotExist` is raised. However the exception catch was too narrow (only caught `Document.DoesNotExist` but need to catch `SubmissionDocument.DoesNotExist` too). Note we still want to distinguish which type of document does not exist.

Existing behaviour has resulted the submission getting into an odd state. If a document to be replaced does not exist then there is no need to then further raise an API Exception.