feat: TRSV2-32 New Models
This work implements a new authentication app that should be used to replace the existing authentication logic in the API. The implementation is switched off with the API_V2_ENABLED setting (defaults to False) and has no effect on the behaviour of existing authentication code. Test coverage of this package is 100%. Tests are all in trade_remedies_api/tests. Run tests with something like:
API_V2_ENABLED=true pytest trade_remedies_api/tests/