Beta version of the new public file page.
Added ExportSourceSerializer with country full name field

Removed the users field from the ORganisationSerializer as it was costly and not used anywhere

Added size_in_kb to DocumentSerializer.

Refactored the get_public_file CaseViewSet endpoint so that it technically uses the V1 submissions manager method to retrieve the submissions open for the public. Felt it was safer this way and the code is not terrible.